Rework inspection code to move more functionality into the inspect/ package
